Adds new Elastic class, which inherits from AngleEnergy. ElasticSVT and ElasticDBRC now inherit from Elastic.
