+++ CMakeLists.txt
@@ -43,6 +43,7 @@ set(PNDL_SOURCE_LIST src/element.cpp
                      src/multiple_distribution.cpp
                      src/cm_distribution.cpp
                      src/svt.cpp
+                     src/elastic.cpp
                      src/elastic_svt.cpp
                      src/elastic_dbrc.cpp
                      src/energy_grid.cpp
