+++ src/python/pyPapillonNDL.cpp
@@ -56,6 +56,7 @@ extern void init_Watt(py::module&);
 extern void init_AngleEnergyPacket(py::module&);
 extern void init_AngleEnergy(py::module&);
 extern void init_Uncorrelated(py::module&);
+extern void init_Elastic(py::module&);
 extern void init_ElasticSVT(py::module&);
 extern void init_ElasticDBRC(py::module&);
 extern void init_NBody(py::module&);
@@ -121,6 +122,7 @@ PYBIND11_MODULE(pyPapillonNDL, m) {
   init_AngleEnergyPacket(m);
   init_AngleEnergy(m);
   init_Uncorrelated(m);
+  init_Elastic(m);
   init_ElasticSVT(m);
   init_ElasticDBRC(m);
   init_NBody(m);
